+++ themes-default/slim/src/components/manage/index.js
@@ -1,0 +1,2 @@
+export { default as ChangeIndexerRow } from './change-indexer-row.vue';
+export { default as SelectIndexer } from './select-indexer.vue';