+++ themes-default/slim/src/router/routes.js
@@ -343,6 +343,17 @@ const manageRoutes = [
         props: true
     },
     {
+        path: '/manage/changeIndexer',
+        name: 'manageChangeIndexer',
+        meta: {
+            title: 'Change show indexer',
+            header: 'Change show indexer',
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/change-indexer.vue')
+    },
+    {
         path: '/manage/backlogOverview',
         name: 'manageBacklogOverview',
         meta: {