+++ themes-default/slim/src/components/subtitle-search.vue
@@ -213,7 +213,7 @@ export default {
             this.displayQuestion = false;
             this.loading = true;
             this.loadingMessage = 'Searching for subtitles... ';
-            apiRoute('home/manualSearchSubtitles', { params: subtitleParams })
+            apiRoute('home/manualSearchSubtitles', { params: subtitleParams, timeout: 120000 })
                 .then(response => {
                     if (response.data.result === 'success') {
                         this.subtitles.push(...response.data.subtitles);
@@ -263,7 +263,7 @@ export default {
             this.loadingMessage = 'downloading subtitle... ';
             this.loading = true;
 
-            apiRoute('home/manualSearchSubtitles', { params })
+            apiRoute('home/manualSearchSubtitles', { params, timeout: 120000 })
                 .then(response => {
                     if (response.data.result === 'success') {
                         // Update the show, as we have new information (subtitles)