+++ CHANGELOG.md
@@ -3,11 +3,14 @@
 #### New Features
 - Add support for banner and background images to indexer tvmaze ([10234](https://github.com/pymedusa/Medusa/pull/10234))
 - Add option for using ffprobe to validate postprocessed media ([10132](https://github.com/pymedusa/Medusa/pull/10132))
+- Add change indexer page to change the current indexer for shows in bulk ([9862](https://github.com/pymedusa/Medusa/pull/9862))
 
 #### Improvements
 - Add column sorting for the add new show page search results ([10217](https://github.com/pymedusa/Medusa/pull/10217))
 - Add series start year as a renaming option ([10183](https://github.com/pymedusa/Medusa/pull/10183))
 - Remove git username/password authentication. No longer supported by github. ([10144](https://github.com/pymedusa/Medusa/pull/10144))
+- Add option to allow for overwriting nfo files. ([10237](https://github.com/pymedusa/Medusa/pull/10237))
+- Improve kodi nfo file creation. ([10237](https://github.com/pymedusa/Medusa/pull/10237))
 
 #### Fixes
 - Fix displayShow search subtitle button ([10214](https://github.com/pymedusa/Medusa/pull/10214))